modify for correct i/o

+++ Resources/1AbstractProduct/IOrbit.cs
@@ -7,6 +7,6 @@ namespace Resources
     public interface IOrbit 
     {
         double Distance();
-        int TotalCraters();
+        double TotalCraters();
     }
 }
